Fix formatting in canvas settings card
