+++ lib/presentation/widgets/settings/settings_general_card.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import 'settings_toggle_tile.dart';
+
 class SettingsGeneralCard extends StatelessWidget {
   const SettingsGeneralCard({
     super.key,
@@ -22,7 +24,7 @@ class SettingsGeneralCard extends StatelessWidget {
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,
           children: [
-            _SwitchSetting(
+            SettingsToggleTile(
               title: 'Auto Save',
               subtitle: 'Automatically save changes',
               value: autoSave,
@@ -30,7 +32,7 @@ class SettingsGeneralCard extends StatelessWidget {
               switchKey: const ValueKey('settings_auto_save_switch'),
             ),
             const SizedBox(height: 16),
-            _SwitchSetting(
+            SettingsToggleTile(
               title: 'Show Tooltips',
               subtitle: 'Display helpful tooltips',
               value: showTooltips,
@@ -40,51 +42,6 @@ class SettingsGeneralCard extends StatelessWidget {
           ],
         ),
       ),
-    );
-  }
-}
-
-class _SwitchSetting extends StatelessWidget {
-  const _SwitchSetting({
-    required this.title,
-    required this.subtitle,
-    required this.value,
-    required this.onChanged,
-    this.switchKey,
-  });
-
-  final String title;
-  final String subtitle;
-  final bool value;
-  final ValueChanged<bool> onChanged;
-  final Key? switchKey;
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        Expanded(
-          child: Column(
-            crossAxisAlignment: CrossAxisAlignment.start,
-            children: [
-              Text(
-                title,
-                style: const TextStyle(fontWeight: FontWeight.bold),
-              ),
-              const SizedBox(height: 4),
-              Text(
-                subtitle,
-                style: Theme.of(context).textTheme.bodySmall,
-              ),
-            ],
-          ),
-        ),
-        Switch(
-          key: switchKey,
-          value: value,
-          onChanged: onChanged,
-        ),
-      ],
     );
   }
 }
